[feature][basic-navigation] - Implemented complete MouseLook
