--- Assets/Scripts/MouseLook.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b25368711a2257b44be64e3b9dc03909
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
